+++ src/main/java/com/example/demo/model/repository/CustomerRepository.java
@@ -1,25 +1,8 @@
 package com.example.demo.model.repository;
 
 import com.example.demo.model.entity.Customer;
-import org.springframework.stereotype.Repository;
+import org.springframework.data.jpa.repository.JpaRepository;
 
-import javax.persistence.EntityManager;
-import java.util.List;
-
-@Repository
-public class CustomerRepository {
-    private final EntityManager entityManager;
-
-    public CustomerRepository(EntityManager entityManager) {
-        this.entityManager = entityManager;
-    }
-
-    public Customer findById(int id) {
-        return entityManager.find(Customer.class, id);
-    }
-
-    public List<Customer> findAll() {
-        return entityManager.createNativeQuery("SELECT * FROM customers", Customer.class).getResultList();
-    }
+public interface CustomerRepository extends JpaRepository<Customer, Integer> {
 
 }
