+++ src/main/java/com/example/demo/model/repository/ProductRepository.java
@@ -1,25 +1,7 @@
 package com.example.demo.model.repository;
 
 import com.example.demo.model.entity.Product;
-import org.springframework.stereotype.Repository;
+import org.springframework.data.jpa.repository.JpaRepository;
 
-import javax.persistence.EntityManager;
-import java.util.List;
-
-@Repository
-public class ProductRepository {
-    private final EntityManager entityManager;
-
-    public ProductRepository(EntityManager entityManager) {
-        this.entityManager = entityManager;
-    }
-
-    public Product findById(int id) {
-        return entityManager.find(Product.class, id);
-    }
-
-    public List<Product> findAll() {
-        return entityManager.createNativeQuery("SELECT * FROM products", Product.class).getResultList();
-    }
-
+public interface ProductRepository extends JpaRepository<Product, Integer> {
 }
