+++ src/main/java/com/example/demo/service/CustomerService.java
@@ -1,0 +1,13 @@
+package com.example.demo.service;
+
+import com.example.demo.model.entity.Customer;
+
+import java.util.List;
+
+public interface CustomerService {
+
+    List<Customer> findAll();
+
+    Customer findById(int id);
+
+}
