+++ src/main/java/com/example/demo/service/ProductService.java
@@ -1,0 +1,12 @@
+package com.example.demo.service;
+
+import com.example.demo.model.entity.Product;
+
+import java.util.List;
+
+public interface ProductService {
+
+    List<Product> findAll();
+
+    Product findById(int id);
+}
