+++ src/main/resources/application.properties
@@ -1,1 +1,8 @@
-spring.profiles.active=java
+spring.datasource.url=jdbc:h2:mem:spring-jpa
+spring.datasource.driver-class-name=org.h2.Driver
+spring.datasource.username=sa
+spring.datasource.password=
+spring.jpa.hibernate.ddl-auto=none
+spring.jpa.show-sql=true
+spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.H2Dialect
+spring.main.web-application-type=none
